mcp-server/requirements.txt
# FastMCP Framework
fastmcp==0.1.0

# Financial Data & APIs
yfinance>=0.2.28
requests>=2.31.0
pandas>=2.0.0
newsapi-python>=0.2.7

# Technical Analysis
pandas-ta>=0.3.14b0

# Data Processing
numpy>=1.24.0

# Environment & Config
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.8.0
asyncio-contextmanager>=1.0.0

# Utilities
pytz>=2023.3
python-dateutil>=2.8.2

# Optional: NSE Data (for Indian stocks)
nsepython>=0.1.0  # Uncomment to enable NSE integration

# Logging & Monitoring
structlog>=23.1.0

# Type Checking (Development)
mypy>=1.0.0
types-requests>=2.28.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
